Instructions for OPENAI_API_KEY as env var and in IDEs
I've seen a bunch of folks (including myself) accidentally paste their API keys into PRs. I think this is because many are unaware that you can just add an .env file with your key. Adding that as a pointer in README.md - while I'm not sure if this is the place it should go, I think a pointer like this would be helpful.